fix(crawlers): move discover.js handler tests out of app/api/src/routes/
omadaDiscover.test.js and the new entraDiscover.test.js both lived in
app/api/src/routes/ alongside jobs.js — violating the "nothing
crawler-specific belongs outside its tools/crawlers/<type>/ folder" rule
that everything else (ConfigWizard.jsx, Summary.jsx, CrawlerMeta.js,
discover.js itself, and every other test type) already follows.

Moves both to tools/crawlers/<type>/discover.test.js, importing the handler
via a plain './discover.js' relative path. Adds app/api/vitest.config.js
with a test.include glob ('../../tools/crawlers/**/discover.test.js') so
they keep running under `npm test` in app/api — without it they'd silently
stop executing, since app/api has no prior vitest config and its default
test discovery never looks outside its own directory tree (confirmed
empirically: a probe test dropped into tools/crawlers was not picked up
before this config existed).

Updates tools/crawlers/CLAUDE.md, app/api/CLAUDE.md, and
docs/sync/building-a-crawler.md to document the corrected convention and
prevent this drift from recurring; fixes a stale path reference in
jobs.omada.test.js's header comment.
